requirementsUnix.txt
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.7.1
Automat==0.7.0
channels==2.1.2
constantly==15.1.0
coverage==4.5.1
daphne==2.2.1
Django==2.1
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
psycopg2==2.7.5
PyHamcrest==1.9.0
pytz==2018.5
six==1.11.0
Twisted==18.7.0
txaio==18.7.1
zope.interface==4.5.0